+++ Leomard/Models/Lemmy/Forms/ModTools/RemoveComment.swift
@@ -1,0 +1,14 @@
+//
+//  RemoveComment.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 27/07/2023.
+//
+
+import Foundation
+
+struct RemoveComment: Codable {
+    let commentId: Int
+    let reason: String?
+    let removed: Bool
+}